requirements.txt
fastapi
sqlalchemy
jwt
passlib
python-jose